@@ -12,11 +12,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from copy import deepcopy
+
+import numpy as np
+import pytest
+
 from mlip.simulation.enums import SimulationType
 from mlip.simulation.jax_md.jax_md_simulation_engine import JaxMDSimulationEngine
 
 
-def test_minimization_can_be_run_with_jax_md_backend(setup_system_and_mace_model):
+@pytest.mark.parametrize("is_batched", [True, False])
+def test_minimization_can_be_run_with_jax_md_backend(
+    setup_system_and_mace_model, is_batched
+):
     atoms, _, _, mace_ff = setup_system_and_mace_model
 
     md_config = JaxMDSimulationEngine.Config(
@@ -36,7 +44,13 @@ def test_minimization_can_be_run_with_jax_md_backend(setup_system_and_mace_model
         assert state.temperature is None
         assert state.forces is not None
 
-    engine = JaxMDSimulationEngine(atoms, mace_ff, md_config)
+    if is_batched:
+        engine = JaxMDSimulationEngine(
+            [atoms, deepcopy(atoms), deepcopy(atoms)], mace_ff, md_config
+        )
+    else:
+        engine = JaxMDSimulationEngine(atoms, mace_ff, md_config)
+
     engine.attach_logger(_mock_logger)
 
     engine.run()
@@ -45,7 +59,34 @@ def test_minimization_can_be_run_with_jax_md_backend(setup_system_and_mace_model
     assert engine.state.compute_time_seconds > 0.0
     assert engine.state.temperature is None
     assert engine.state.kinetic_energy is None
-    assert engine.state.positions.shape == (10, 10, 3)
-    assert engine.state.forces.shape == (10, 10, 3)
     assert engine.state.velocities is None
     assert intermediate_steps == [10, 20]
+
+    expected_first_atom_forces = np.array([0.0498, -0.0216, 0.0118])
+
+    if is_batched:
+        assert len(engine.state.forces) == 3
+        assert len(engine.state.positions) == 3
+
+        for i in range(3):
+            assert engine.state.forces[i].shape == (10, 10, 3)
+            assert engine.state.positions[i].shape == (10, 10, 3)
+
+        assert np.allclose(engine.state.forces[0], engine.state.forces[1], atol=1e-3)
+        assert np.allclose(
+            engine.state.positions[0], engine.state.positions[1], atol=1e-3
+        )
+        assert np.allclose(engine.state.forces[0], engine.state.forces[2], atol=1e-3)
+        assert np.allclose(
+            engine.state.positions[0], engine.state.positions[2], atol=1e-3
+        )
+        assert np.allclose(
+            engine.state.forces[0][0][0], expected_first_atom_forces, atol=1e-3
+        )
+
+    else:
+        assert engine.state.forces.shape == (10, 10, 3)
+        assert engine.state.positions.shape == (10, 10, 3)
+        assert np.allclose(
+            engine.state.forces[0][0], expected_first_atom_forces, atol=1e-3
+        )
